correção e tratamento de erros, reajuste na interface

+++ app/atracoes/deletar/[id]/pageClient.tsx
@@ -11,6 +11,15 @@ export default function DeletarAtracaoPageClient( { id }: { id: string } ) {
     async function deletarAtracao ( e: any ) {
         e.preventDefault();
 
+        const res = await fetch(`/api/atracoes/${id}`);
+        const atracao = await res.json();
+
+        if (atracao.obras.length !== 0) {
+            alert("Não é possivel deletar uma atração que possui obras vinculadas");
+
+            return;
+        }
+
         await fetch(`/api/atracoes/${id}`, {
             method: "DELETE",
         });
@@ -22,10 +31,10 @@ export default function DeletarAtracaoPageClient( { id }: { id: string } ) {
 
     return(
         <div className="max-w-xl mx-auto p-6 flex gap-4">
-             <button type='button' onClick={deletarAtracao} className={`${styles.navButton}`}>Deletar atração</button>
-             <button type='button' onClick={() => {
+            <button type='button' onClick={() => {
                     router.push("/atracoes")
                 }} className={`${styles.navButton}`}>Voltar à atrações</button>
+            <button type='button' onClick={deletarAtracao} className={`${styles.navButton}`}>Deletar atração</button>
         </div>
     );
